Optimize
1.
Create a class, Triangle. Its __init__() method should take self, angle1, angle2, and angle3 as arguments. 
Make sure to set these appropriately in the body of the __init__()method.
Create a variable named number_of_sides and set it equal to 3.
Create a method named check_angles. The sum of a triangle's three angles should return True if the sum of self.angle1, self.angle2, and self.angle3 is 
equal 180, and False otherwise.
Create a variable named my_triangle and set it equal to a new instance of your Triangle class. Pass it three angles that sum to 180 (e.g. 90, 30, 60).
Print out my_triangle.number_of_sides and print out my_triangle.check_angles().

2.Create a class called Employee whose objects are records for an employee.
This class will be a derived class of the class Person .
An employee record has an employee's name and other details like age,place and mob(inherited from the class Person), an annual salary represented as a single value of type double, a year the employee started work as a single value of type int and a national insurance number, which is a value of type String.
Your class should have a reasonable number of constructors and accessor methods, as well as an equals method. 
Class Employee should have a method that calculates the salary to be debited based on the number of days he/she was present.Assuming 3460 is the pay per day for an employee.